dropship-showcase/backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-ratelimit==4.1.0
psycopg2-binary==2.9.11
python-decouple==3.8
dj-database-url==3.1.2
gunicorn==23.0.0
